Integrate DataTables Library
I added datatables.net and datatables.net-dt to the project as Bower dependencies. These versions will be used for testing with the dummy application and automated tests.

I created the default blueprint and added instructions to install datatables.net and datatables.net-dt as Bower packages for applications that install the ember-datatables addon.

I updated index.js to include the Bower JavaScript, CSS, and image assets in the generated Ember application.