Implement data-table Component
I added the data and columns properties to the data-table component. The data property will accept the list of Ember Data records such as the model for a route to be displayed in the table. The columns property contains the title of the table column and the name of the property to be displayed in the column cell.

I installed ember-cli-mirage. Mirage is being used to simulate a web API server to load the test records to be displayed in the dummy application's tables.

I created the employee model to represent test data that I am displaying in the tests for the data-table component.