Create datetime-renderer Helper
I created the datetime-renderer Ember helper that wraps the datetime renderer plugin for DataTables in order to render dates and times properly in a table.

I added ember-moment as a dependency for the ember-datatables addon.

I added datatables.net-plugins as a bower dependency for the ember-datatables package.

I updated the dummy application. I added more sample data and updated the table to use the datetime-renderer helper to format the employee start date.